lowered framecap (60 -> 30) so that experience is not hampered for low end devices
unless the device runs it at less than 30 fps, in which case there's not much i can do. If I lowered the framecap even further it'd just start to look terrible for everyone

sidenote: now that fps is capped, the animation will pause when out of view. nice by-product, totally meant it, yep